requirements.txt
streamlit==1.32.0
rembg==2.0.65
Pillow==10.1.0
numpy==1.26.0 